requirements.txt
sentry-sdk
pandas
python-dotenv
redlock-py
geoip2